requirements.txt
aiogram==3.5.0
python-dotenv==1.0.1
openai==1.25.1